requirements.txt
# Backend dependencies
fastapi>=0.115.0
uvicorn>=0.34.0
google-generativeai>=0.8.0
pydantic>=2.0.0

# Frontend dependencies
streamlit>=1.43.0
requests>=2.30.0

# Common dependencies
python-dotenv>=1.0.0 